+++ example/lib/routes.dart
@@ -1,3 +1,4 @@
+import 'package:example/homepage/lottie_index.dart';
 import 'package:example/homepage/network_tools/bili_favorite_video/bili_favorite_video_binding.dart';
 import 'package:example/homepage/network_tools/bili_favorite_video/bili_favorite_video_view.dart';
 import 'package:example/homepage/network_tools/bili_hot_video/bili_hot_video_binding.dart';
@@ -7,6 +8,7 @@ import 'package:example/homepage/network_tools/history_weather_query/history_wea
 import 'package:example/homepage/network_tools/history_weather_query/history_weather_query_view.dart';
 import 'package:example/homepage/simple_information/device_info/device_info_binding.dart';
 import 'package:example/homepage/simple_information/device_info/device_info_view.dart';
+import 'package:example/homepage/text_index.dart';
 import 'package:redstonex/routes/dispatcher.dart';
 import 'package:redstonex/routes/router.dart';
 
@@ -28,6 +30,14 @@ class Routes {
         .newRoute(Router(RouteCompose.biliRoute.videoDetail, () => BiliWebView()))
         .newRoute(Router(RouteCompose.biliRoute.favoriteVideos, () => BiliFavoriteVideoPage(),
             binding: BiliFavoriteVideoBinding()));
+
+    RouterGroup lottieGroup = XDispatcher.group(groupName: 'lottie');
+    lottieGroup.newRoute(
+        Router(RouteCompose.lottieRoute.index, () => LottieIndex(),));
+
+    RouterGroup textGroup = XDispatcher.group(groupName: 'text');
+    textGroup.newRoute(
+        Router(RouteCompose.textRoute.index, () => TextIndex(),));
   }
 }
 
@@ -36,6 +46,8 @@ class RouteCompose {
   static DeviceRoute deviceRoute = DeviceRoute();
   static WeatherRoute weatherRoute = WeatherRoute();
   static BiliRoute biliRoute = BiliRoute();
+  static LottieRoute lottieRoute = LottieRoute();
+  static TextRoute textRoute = TextRoute();
 }
 
 class DeviceRoute {
@@ -51,3 +63,12 @@ class BiliRoute {
   String videoDetail = '/bili/hotVideo/detail';
   String favoriteVideos = '/bili/favorite/home';
 }
+
+class LottieRoute {
+  String index = '/lottie/home';
+}
+
+class TextRoute {
+  String index = '/text/home';
+}
+