+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.2
+* 增加支持Lottie动画的Bottom Navigation Bar-PageView
+* 增加persistent_bottom_nav_bar_v2三方库的文字Bottom Navigation Bar-PageView
+
 ## 1.1.1
 * 增加 `<String, dynamic>` 类型map扩展方法